+++ views/homepage/index.ejs
@@ -31,7 +31,7 @@
           <div class="sign-in">
             <form
 
-            method="post"
+            method="POST"
             action="/api/login"
             class="form-inline"
             id="log-in"
@@ -41,7 +41,7 @@
             <div class="form-group">
               <input
               type="email"
-              name="user[email]"
+              name="email"
               class="form-control"
               id="email"
               placeholder="Email"
@@ -55,7 +55,7 @@
               <input
 
               type="password"
-              name="user[password]"
+              name="password"
               class="form-control"
               placeholder="Password"
               required=""
